shared: add NMIPAddr struct
