+++ src/nm-core-utils.h
@@ -90,25 +90,6 @@ GETTER (void) \
 
 /*****************************************************************************/
 
-typedef struct {
-	union {
-		guint8 addr_ptr[1];
-		in_addr_t addr4;
-		struct in6_addr addr6;
-
-		/* NMIPAddr is really a union for IP addresses.
-		 * However, as ethernet addresses fit in here nicely, use
-		 * it also for an ethernet MAC address. */
-		guint8 addr_eth[6 /*ETH_ALEN*/];
-	};
-} NMIPAddr;
-
-extern const NMIPAddr nm_ip_addr_zero;
-
-#define NMIPAddrInit { .addr6 = IN6ADDR_ANY_INIT }
-
-/*****************************************************************************/
-
 guint nm_utils_in6_addr_hash (const struct in6_addr *addr);
 
 gboolean nm_ethernet_address_is_valid (gconstpointer addr, gssize len);
